requirements.txt
gradio
torch
torchvision
ultralytics
Pillow